Extracted common CC and non-CC performance report methods into a separate module
